value check added

+++ src/main/java/com/practice/qtrproject/service/WriterService.java
@@ -6,7 +6,6 @@ import com.practice.qtrproject.mapper.WriterMapper;
 import com.practice.qtrproject.model.writer.Writer;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.ibatis.annotations.Mapper;
 import org.springframework.beans.BeanUtils;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -34,7 +33,11 @@ public class WriterService {
     }
 
     @Transactional(rollbackFor = Exception.class)
-    public void updateWriterInfo(ReqWriterInfoUpdDto updDto) {
+    public void updateWriterInfo(ReqWriterInfoUpdDto updDto) throws Exception {
+        int valueCount = writerMapper.checkVal(updDto.getWriterNo());
+        if(valueCount == 0){
+            throw new Exception("no value");
+        }
         try {
             writerMapper.updateInfo(updDto);
         }catch (Exception e){
@@ -44,11 +47,16 @@ public class WriterService {
 
 
     @Transactional(rollbackFor = Exception.class)
-    public void deleteWriter(Integer writerNo){
+    public void deleteWriter(Integer writerNo) throws Exception {
+        int valueCount = writerMapper.checkVal(writerNo);
+        if(valueCount == 0){
+            throw new Exception("no value");
+        }
         try {
             writerMapper.deleteWriterInfo(writerNo);
         }catch (Exception e){
             log.error("error on deleting : {}", e.getMessage());
+            throw new Exception(e.getMessage());
         }
     }
